edited checkboxes behavior
