+++ src/components/flights/add-flight-popup/material.style.js
@@ -1,6 +1,6 @@
 const styles = () => ({
   dialogButton: {
-    backgroundColor: '#efefef',
+    backgroundColor: 'white',
     '&:hover': {
       backgroundColor: '#dbdbdb'
     }
