+++ src/components/flights/index.js
@@ -1,16 +1,14 @@
-import React, { Component } from 'react';
+import React from 'react';
 import AddFlightPopup from './add-flight-popup';
-import FlightsList from './flights-list/material.table';
+import FlightsList from './flights-list';
 
-class FlightsPage extends Component {
-  render() {
-    return (
-      <section className="flights-page">
-        <AddFlightPopup />
-        <FlightsList />
-      </section>
-    )
-  }
+function FlightsPage() {
+  return (
+    <section className="flights-page">
+      <AddFlightPopup />
+      <FlightsList />
+    </section>
+  );
 }
 
 export default FlightsPage;
